+++ jeap-error-handling-service/src/test/java/ch/admin/bit/jeap/errorhandling/domain/eventhandler/ErrorEventMapperTest.java
@@ -1,4 +1,4 @@
-package ch.admin.bit.jeap.errorhandling.domain.eventHandler;
+package ch.admin.bit.jeap.errorhandling.domain.eventhandler;
 
 import ch.admin.bit.jeap.errorhandling.infrastructure.kafka.DomainEventDeserializerProvider;
 import ch.admin.bit.jeap.errorhandling.infrastructure.persistence.OriginalTraceContext;